+++ CHANGELOG.md
@@ -1,12 +1,12 @@
-* migrated Android Plugin to new Flutter 1.12 API
-
 ## Upcoming release
+* migrated Android Plugin to new Flutter 1.12 API
 * fix fetching device dnd for fresh installations
 * fix custom params & headers on cloud API requests
 * fix beacons parsing when the result is empty
 * added method to send a password reset email
 * fix user mapping
 * enhance NotificareUser model
+* fix handling new intents
 
 ## 2.2.4
 * check partially fetched notifications when fetching inbox items